This adds basic visualization capabilities.
This adds a module that shows how the vispy
library can be used to generate visual effects.
The goal is to make it possible for output from the
touch panel to create interesting visual effects when
the user touches it.

The code added still needs to be hooked up to the rest
of the application so that it sets the color of the
visual element based on the note that is playing and
the location of the visualization based on the point of
contact on the touch panel.